pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "nanobot-cloud"
version = "0.1.0"
description = "Nanobot AI Assistant - Cloud Deployment"
authors = [{name = "Julian Juarez (QUINTANA)", email = "dev@quintana.local"}]
requires-python = ">=3.11"

[tool.poetry]
name = "nanobot-cloud"
version = "0.1.0"
description = "Nanobot AI Assistant - Cloud Deployment with Telegram, MCP, Dashboard"
authors = ["Julian Juarez <dev@quintana.local>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.111.0"
uvicorn = {version = ">=0.30.0", extras = ["standard"]}
python-telegram-bot = ">=21.0"
anthropic = ">=0.34.0"
groq = ">=0.11.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
boto3 = "^1.29.0"
httpx = ">=0.27.0"
aiofiles = "^23.2.1"
tenacity = "^8.2.3"
loguru = "^0.7.2"
mcp = ">=0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.8"
mypy = "^1.7.0"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
